Consume BOPTEST from a subtree
In the previous commit, the canonical BOPTEST dependency was added as a
git subtree. In this commit, the pieces of BOPTEST (unrelated to BOPTEST-service),
which have been made redundant by the subtree are removed.

At this stage BOPTEST-Service is functional while consuming BOPTEST in
the new subtree location. The BOPTEST tests still need to be updated in
order to work with the service because the subtree and therefore the
tests are based on the BOPTEST master branch. A future commit needs to
reintroduce the BOPTEST-Service flavor of the tests.

All of this is an effort to retain history while still migrating
BOPTEST-Service to a dedicated repository and introducing BOPTEST
(canonical) as a subtree.