Remove validation of testid
Validating the testid has become more expensive now that tests can be
associated with a user (which requires a trip to validate the user),
therefore testid validation "in the loop" of simulation apis has been
removed. Now instead of getting an immediate error, the response will
timeout after not receiving a response from a running test.