Align pub/sub message format with canonical BOPTEST api doc
Previously there was some light transformation of the data sent over
pub/sub, resulting in a slight difference between the api payloads documented
on the IBPSA BOPTEST repo and the expected pub/sub payloads.

This has not been an issue in the past because the differences were
internal, and the rest API of BOPTEST-Service was still aligned with the
canonical BOPTEST. With the introduction of a websocket api, clients are
given access to directly put messages on the pub/sub bus, so for the
sake of clarity it is good to have alignment.